+++ wpull/builder.py
@@ -535,7 +535,11 @@ class Builder(object):
 
     def _build_html_parser(self):
         '''Build HTML parser.'''
-        from wpull.document.htmlparse.lxml_ import HTMLParser
+        if self._args.html_parser == 'html5lib':
+            from wpull.document.htmlparse.html5lib_ import HTMLParser
+        else:
+            from wpull.document.htmlparse.lxml_ import HTMLParser
+
         self._factory.class_map['HTMLParser'] = HTMLParser
         self._factory.new('HTMLParser')