+++ doc/changelog.rst
@@ -5,6 +5,7 @@ What's New
 * Fixed ``--warc-move`` option which had no effect.
 * Support for HTTP/HTTPS proxies but no HTTPS tunnelling support. Wpull will refuse to start without the insecure override option. Note that if authentication and WARC file is enabled, the username and password is recorded into the WARC file.
 * Added ``--ignore-fatal-errors`` option.
+* Added ``--http-parser`` option. You can now use html5lib as the HTML parser.
 
 * API: